Isaiah 5:7 change 2nd person to 3rd
Isaiah 5:7 changed the 2nd person pronouns to 3rd in lines 3 & 5 since this is not consistent with the ULB, the Hebrew or any English version I have seen. In fact, personal pronouns are not used at all but 3rd person is better than 2nd since he is not addressing the people, only talking ABOUT them.