Matthew 9:5 wording change
Matthew 9:5 Original wording:
It would not be hard for someone to say to this man, 'I forgive your sins,' because no one could see whether or not it really happened. But no one would say to him, 'Get up and walk!,' because people could easily see whether he could walk again!
Changed because this loses the rhetorical question and reverses the idea that it IS HARDER to forgive sins than to heal the paralytic.