Matthew 8:4 simplify and condense
Matthew 8:4 simplify and condense the first two sentences of the quoted portion: "Go and show yourself to a nearby priest so that he can examine you and verify that you are no longer sick. Then, after he tells the local people, they will know that you no longer have leprosy, and you will be able to associate with them again." since this is background information as is in the notes. The following also was cut, "that people who have been healed from this sickness should offer as a sacrifice to God." and changed the wording to reflect the same idea.