Matthew 9:6 eliminate added phrase
Matthew 9:6 eliminate added phrase "of people while I am on the earth." This is not in the original, is not needed for the sense and is likely not even theologically correct since we assume Jesus continues to forgive sins (1 John 2). Besides, the noun "sins" is anarthrous so is a generalized term.