Matthew 9:10 eliminate extra phrase
Matthew 9:10 eliminate extra clause "who did not care about the law of Moses" is not certain. We surmise that they considered themselves beyond any hope of salvation since their track record spoke against them in the theology of the Pharisees, so they settled into a despair. Jesus' message of salvation was the magnet that drew them. Regardless of their spiritual state, to say they did not care is saying too much.