Matthew 9:21 eliminate extra wording
Matthew 9:21 eliminate extra wording "She was saying to herself, 'I want Jesus to heal me without anyone finding out that I have this problem of bleeding. So if I touch him or even if I just touch his garment, I will be healed without anyone finding out about it.' So" since this is explained in the notes and is not in the text.