Matthew 9:21 changed versification
Matthew 9:21 changed versification to reflect the original and the portion eliminated from the previous versification 20-21. There was too much assumption in the wording about the widow's thought process. Since the notes also cover this, there is no need for this in this rendition and could be confusing for the MTTs.