Matthew 12:17 change wording
Matthew 12:17 change wording " acting humbly like that" to what you see. To say that Jesus was humble could be surmised from the Messianic Secret, but that is an assumption and is not at all supported by the quote from Isaiah that stresses the Servant's kindness from the point of victory and exaltation.