Matthew 12:49-50 change wording
Matthew 12:49-50 change wording from, "I love as much as I love my mother and brothers." to what you see. The elided application defends Jesus' physical relationship as much as his spiritual one. But this is not what he says. He replaces the physical with the spiritual relationship as the ULB shows along with the original.