Matthew 19:1 change wording
Matthew 19:1 change wording from ""In order to illustrate how God rewards people, I will compare the way God begins to rule his people with what the owner of an estate did." to what you see. The wording did not reflect the original at all and tended to confuse the MTTs. And the concept of the kingdom is changed.