Matthew 22:16 cut out a lot
Matthew 22:16 cut out a lot of the material already in the notes that serves here as background information and is not in the narrative text. ", who thought that the Israelites should pay only the tax that the Jewish authorities required them to pay. They also sent some members of the party that supported King Herod. The members of that party thought that the Israelites should pay only the tax that the Roman government required them to pay."
Also, "even if it is an important person who does not like what you teach" to what you see.