Mt 23:13 changed wording about kingdom
Mt 23:13 changed wording about kingdom from: "because you keep people from accepting him to rule over them." to what you see. I restored the word "heaven" since this is an important concept in Matthew that should not be lost. The rule of heaven includes Jesus, the Christ, as ruler and as God which is part of Matthew's distinct message compared to the other gospels.