Mt 23:13 cut redundant wording
Mt 23:13 cut redundant wording in the last two sentences. This did read: "You are like men who close a gate so that others cannot enter God's city. You yourselves do not want to go in, and you keep others from entering, too." I cut both sentences because this added a simile not in the original and the redundancy could confuse the MTTs.