Mt 24:30 cut extra wording and clarify
Mt 24:30 cut extra wording and clarify from: "everyone will see something in the sky that indicates that I, the Son of Man, am returning to the earth." to what you see. The original wording misled by "see something that indicates" instead of following the textual idea that the Son himself will appear.