Mt 24 the whole chapter. Remove personal references
Mt 24 the whole chapter. Remove personal references except where Jesus talks about what he says. All the pronominal references to Jesus as the Son of Man and the coming one are interpretive and beyond the text. Jesus is referring to himself but only the third person. I also cut out much of the redundancy. This became too much to explain in detail.