Matthew 25 remove al first person references
Matthew 25 remove all first person references to Jesus as the coming Messiah in the apocalyptic discourse as in chapter 24. These are interpretive (even if correct!) and not in the text and would confuse MTTs who rely on this. The notes should provide the explanation as they do when they say that Jesus speaks about himself in the third person.