Mt 3:2 change wording re kingdom
Mt 3:2 change wording re kingdom from, "God will soon show that he is king" to what you see. This and other instances are changed to reflect Matthew's important distinction of using "heaven" rather than "of God" because the rule includes Jesus triumphant. This also is more than "showing" that he is king, but rather that he IS king and actively exercising his rule.