Rom 1:14 change wording
Romans 1:14 change the loaded terms "educated and to those who are uneducated" to what you see. There are many MTTs and their readers who may not have much formal education, but still have good discretion and ability to learn. Besides, this does not convey  the idea of "wise" and "foolish" in the text.