Romans 1:27 change wording
Romans 1:27 change wording from, "As a result, God has punished them by sicknesses in their bodies, which is what they deserve." to what you see. To ascribe the agency to God is true ultimately but is too much to read out of the text and is likely offensive to some. The more immediate agency is the physical effects of such behavior that have become obvious in our context as well.