Update 02.usfm
v. 3: "the one who will destroy everything that God has done"
changed to:
"the one whom God will destroy"
This edit translates the ULB "son of destruction" and follows the UBS Notes and the NIV, interpreting the "son of destruction," who is "the man of lawlessness," with the view that this man will not so much "destroy everything God has done," but instead he himself will be destroyed by God.