Ezk 3:12
I have changed ULB and UDB to follow NIV and NLT. ESV and NAS take the
parenthesis to be a quotation from an unnamed speaker, probably one
whose voice is like an earthquake. NET and NRSV take the sound to be
that of Yahweh rising from his place. The alternatives not handled by
U?B are dealt with in the notes.